Use Worker instead of ForegroundService for receiving downloads
