+++ main/build.gradle
@@ -429,6 +429,9 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-common-java8:$lifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-reactivestreams:$lifecycleVersion"
 
+    // workmanager (to replace foreground services for downloader)
+    implementation "androidx.work:work-runtime:2.7.1"
+
     // Support Library AppCompat
     implementation 'androidx.appcompat:appcompat:1.5.1'
